Change most instances of Integer for float. mini2.tsp was not working

+++ src/edu/rowan/algorithms/BruteForceSolver.java
@@ -106,7 +106,7 @@ public class BruteForceSolver {
 
     private double getTotalDistance(ArrayList<Integer> cities) {
         double total = 0.0;
-        int x1, y1, x2, y2;
+        float x1, y1, x2, y2;
 
         for (int i = 0; i < cities.size(); i++) {
             x1 = tour.getXCoord(cities.get(i));
@@ -119,8 +119,8 @@ public class BruteForceSolver {
                 x2 = tour.getXCoord(cities.get(0));
                 y2 = tour.getXCoord(cities.get(0));
             }
-            int[] p = {x1, y1};
-            int[] q = {x2, y2};
+            float[] p = {x1, y1};
+            float[] q = {x2, y2};
 
             total += getDistance(p, q);
         }// end for loop
@@ -129,7 +129,7 @@ public class BruteForceSolver {
     }// end of getTotalDistance()
     
     
-    private double getDistance(int[] p, int[] q) {
+    private double getDistance(float[] p, float[] q) {
         double a = 0.0;
         // dist((x, y), (a, b)) = √(x - a)² + (y - b)²
         a = Math.sqrt((Math.pow((p[0] - q[0]), 2)) + (Math.pow((p[1] - q[1]), 2)));
